Disambiguate java-library variants when consumer does not provide any attributes
Since inception, Gradle has assumed that, by default, consumers of a maven repository require
the _runtime_ variant of the published library.

This change adds a disambiguation rule that encodes this assumption for the case where a java library
is published with variants using Gradle module metadata. This will allow us to migrate to consuming
the new module metadata format by default without breaking a bunch of consumers that depend on this assumption,
declaring no preference for a particular variant.